backend/requirements.txt
fastapi
uvicorn[standard]
pandas
scikit-learn
neo4j
python-dotenv
pydantic
